Now deadline and creationTime are timestamp in long,expTime is only for addGeeft function. Feedback value is double

+++ app/src/main/java/samurai/geeft/android/geeft/interfaces/TaskCallbackBooleanArray.java
@@ -6,5 +6,5 @@ import samurai.geeft.android.geeft.adapters.GeeftItemAdapter;
  * Created by danybr-dev on 08/02/16.
  */
 public interface TaskCallbackBooleanArray {
-        void done(boolean result,GeeftItemAdapter.ViewHolder holder, long[] userInformation);
+        void done(boolean result,GeeftItemAdapter.ViewHolder holder, double[] userInformation);
 }
